% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{addHD}
\alias{addHD}
\title{Add two hyper-dual numbers}
\usage{
addHD(x, y, verbose = FALSE)
}
\arguments{
\item{x}{A vector of length four containg a hyper-dual number}

\item{y}{A vector of length four containg a hyper-dual number}

\item{verbose}{A boolean variable with default \sQuote{false}
indicating if verbose operation is desired.}
}
\value{
A vector of length four containing the resulting
hyper-dual number.
}
\description{
This function adds two hyper-dual numbers
}
\examples{
addHD( c(1,0,0,1), c(2,1,0,-1) )
}
\author{
Dirk Eddelbuettel
}

