\name{RcppHyperDual-package}
\alias{RcppHyperDual-package}
\alias{RcppHyperDual}
\docType{package}
\title{
What the package does (short line)
}
\description{
More about what it does (maybe more than one line)
}
\details{
An overview of how to use the package, including the most important functions
}
\author{
Your Name

Maintainer: Your Name <your@email.com>
}
\references{
}
\keyword{ package }
\examples{
   addHD( c(1,2,0,0), c(2,-1,0,1))
}
