\name{RcppHyperDual-package}
\alias{RcppHyperDual-package}
\alias{RcppHyperDual}
\docType{package}
\title{
  Rcpp Interface to Hyper-Dual Numbers
}
\description{
  This package provides an interface to Hyper-Dual Numbers by Jeffrey Fike and Juan Alonson.
}
\details{
  To be done.
}
\author{
  Dirk Eddelbuettel

  Maintainer: Dirk Eddelbuettel <edd@debian.org>
}
\keyword{ package }
\examples{
   addHL( c(1,2,0,0), c(2,-1,0,1))
}
